Add the `PodOption<T>` type
For [permissioned
markets](https://github.com/openbook-dex/openbook-v2/issues/33), we need
to store the `Pubkey`s of market authorities. Assuming that we do not
want to require *all* markets to be permissioned, we will need to store
both (1) whether or not this market has an authority and (2) the
authority, if any.

Traditionally, this would be done in an `Option<Pubkey>`, but we use
[zero copy
deserialization](https://docs.rs/anchor-lang/latest/anchor_lang/prelude/attr.zero_copy.html),
which requires all types to implement `bytemuck::Pod`. `std::option::Option<T>` does not implement `Pod`, and because of the way enum bit patterns work in Rust, it cannot. So we need to add a custom option implementation that *does* implement `Pod`.